chore: 배너 섹션 폴더로 옮기기
