+++ client/src/Pages/Main/BannerSection/Banner.tsx
